Make all the set buttons save all the settings
Drivers were occasionally trying to enter all the values then clicking
set and it was overwriting the other ones.  So when any of the Set
buttons is clicked just save all the settings.

I tried replacing the three buttons with one big button but I thought it
looked less intuitive that way, so I just left the three.