Add camUrlList
Sir Thorn wanted to have a list of cam locations instead of an open text
field.  If you leave the camUrlList as an empty array in the config in
index.js it will use the open input field.  If you define the array, it
will use a dropdown list containing those items.