playlistDriver stop when no one is listening.
I joined the session in progress the other day but it was way farther
into a file than I was ready to be, so I bailed out and played something
else myself where I could start from the beginning.  I noticed other
users join for a few minutes and then leave also.  So if there are no
riders, just stop.  If one joins, start a new file.  If one leaves and
then immediately re-joins, continue that same file.  After a minute,
start a fresh one.